My proposed changes to the Orders.cpp
Basically this is my implementation of part 4 for A2. The logic is there, just a bit confused about Negotiate order and how to know from where the attack is coming from
and how to write a control flow to deal with that so the Negotiate::execute() is where I am. I have not yet look at the "this order can only be created by this type of Card" yet.

All of the above need test, but feel free to give your opinions on these.